[01100] Consolidate FileAttachmentList into shared component
