+++ src/frontend/src/widgets/inputs/shared/index.ts
@@ -1,0 +1,2 @@
+export { FileAttachmentList } from "./FileAttachmentList";
+export { FileUploadStatus, type FileItem } from "./types";
